fix: stabilize competing chat scroll interactions
